pyproject.toml
[tool.poetry]
name = "stonks"
version = "0.1.0"
description = ""
authors = ["David Shlemayev <david.shlemayev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
xlrd = "^2.0.1"
click = "^8.1.3"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
google-api-python-client = "^2.86.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2"

[tool.poetry.scripts]
stonksd = "stonks.__main__:main"
stonksctl = "stonks.manage:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
